game_news_agent/rawg_kiota_client/pyproject.toml
[project]
name = "rawg-kiota-client"
version = "1.0.0"
description = "RAWG API client generated with Microsoft Kiota"
requires-python = ">=3.13"
dependencies = [
    "microsoft-kiota-abstractions>=1.9.8",
    "microsoft-kiota-http>=1.9.8",
    "microsoft-kiota-serialization-json>=1.9.8",
    "microsoft-kiota-serialization-text>=1.9.8",
    "microsoft-kiota-serialization-form>=1.9.8",
    "microsoft-kiota-serialization-multipart>=1.9.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
